pyproject.toml
[project]
name = "hive-micro"
version = "0.1.6"
description = "A microblogging platform built on Hive blockchain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.16.5",
  "bleach>=6.2.0",
  "flask>=3.1.1",
  "flask-caching>=2.3.1",
  "flask-sqlalchemy>=3.1.1",
  "gunicorn>=23.0.0",
  "hive-nectar>=0.0.7",
  "markdown>=3.8.2",
  "psycopg2-binary>=2.9.10",
  "pygments>=2.18.0",
  "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.12.9",
]
